fix dispatcher on Windows while using thread per connection concurrency model
This patch fixes or better works around issue with Windows select
syscall. The problem is that Windows select signals EINVAL when
rset/wset/xset file desctiptor sets are empty. It simply does not
sleep for the specified amount of time. The work around is to sleep on
our own in such case.