do not terminate GIOPConn threads (danger of cancel/join race-condition)
This fix fixes GIOPConn object to not terminate (pthread_cancel) its
reader/writer threads. Termination is dangerous since it requires
another thread to be already waiting for the thread end (pthread_join)
as we're not using detached threads and when the thread is not
waiting, then "terminate called without an active exception" is
printed to the console and application is aborted. When we allow those
threads to die freely as they are going to do anyway due to changed
state on the GIOPConn object, then no such issue happen.