make code path between fork/exec as short as possible to prevent dead-locks
This patch shortens the code path between fork and exec calls. The
motivation is to avoid any code there especially with the library
calls which may results in memory allocation which may result in
locking of mutexes of memory allocator. This may later in the runtime
results in a hard to debug dead-lock as fork results in copy of whole
memory content to child process including possibly held mutexes, but
it does not copy any threads hence in child there is nothing which may
potentially unlock those locked mutexes.